backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
python-multipart>=0.0.6,<0.1.0

# Database
sqlalchemy[asyncio]>=2.0.0,<2.1.0
alembic>=1.13.0,<1.14.0
asyncpg>=0.29.0,<0.30.0  # PostgreSQL async driver
aiosqlite>=0.19.0,<0.20.0  # SQLite async driver

# Validation & Settings
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Image Processing
pillow>=10.2.0,<11.0.0

# Utilities
python-dateutil>=2.8.0,<3.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0,<0.24.0
httpx>=0.26.0,<0.27.0

# Development
ruff>=0.1.0
black>=24.0.0

